Move common to creation UI pages utils and components to shared directory
