+++ client/web/src/enterprise/insights/components/creation-ui-kit/live-preview-container/LivePreviewContainer.module.scss
@@ -1,7 +1,6 @@
 .live-preview {
     display: flex;
     flex-direction: column;
-    position: relative;
 }
 
 .chart--with-mock {
