+++ client/web/src/enterprise/insights/components/creation-ui-kit/sanitizers/repositories.ts
@@ -1,0 +1,11 @@
+/**
+ * Returns parsed by string repositories list.
+ *
+ * @param rawRepositories - string with repositories split by commas
+ */
+export function getSanitizedRepositories(rawRepositories: string): string[] {
+    return rawRepositories
+        .trim()
+        .split(/\s*,\s*/)
+        .filter(repo => repo)
+}
