+++ client/web/src/enterprise/insights/components/form/hooks/useForm.ts
@@ -42,7 +42,7 @@ interface UseFormProps<FormValues extends object> {
     /**
      * Submit handler for a form element.
      */
-    onSubmit?: (values: FormValues) => SubmissionErrors | Promise<SubmissionErrors> | void
+    onSubmit?: (values: FormValues) => SubmissionResult
 
     /**
      * Change handler will be called every time when some field withing the form
