+++ client/web/src/enterprise/insights/pages/insights/creation/lang-stats/components/lang-stats-insight-creation-form/LangStatsInsightCreationForm.tsx
@@ -64,7 +64,6 @@ export const LangStatsInsightCreationForm: React.FunctionComponent<LangStatsInsi
     const isGqlBackend = api instanceof CodeInsightsGqlBackend
 
     return (
-        // eslint-disable-next-line react/forbid-elements
         <form
             ref={innerRef}
             noValidate={true}
