+++ client/web/src/enterprise/insights/pages/insights/creation/search-insight/components/search-insight-creation-content/validators.ts
@@ -1,12 +1,12 @@
 import { dedupeWhitespace } from '@sourcegraph/shared/src/util/strings'
 
 import { renderError } from '../../../../../../../../components/alerts'
+import { getSanitizedRepositories } from '../../../../../../components/creation-ui-kit/sanitizers/repositories';
 import { Validator } from '../../../../../../components/form/hooks/useField'
 import { AsyncValidator } from '../../../../../../components/form/hooks/utils/use-async-validation'
 import { createRequiredValidator } from '../../../../../../components/form/validators'
 import { fetchRepositories } from '../../../../../../core/backend/core/requests/fetch-repositories'
 import { EditableDataSeries } from '../../types'
-import { getSanitizedRepositories } from '../../utils/insight-sanitizer'
 
 export const repositoriesFieldValidator: Validator<string> = value => {
     if (value !== undefined && dedupeWhitespace(value).trim() === '') {
